+++ authentication_test.go
@@ -42,7 +42,7 @@ func TestAuthenticate(t *testing.T) {
 				fmt.Fprint(w, "{ \"foo\": }")
 			},
 			status: http.StatusOK,
-			err:    "unable to unmarshal JSON into Authentication: invalid character '}' looking for beginning of value",
+			err:    "unable to unmarshal JSON: invalid character '}' looking for beginning of value",
 		},
 		{
 			name: "unauthorized auth",
@@ -56,6 +56,19 @@ func TestAuthenticate(t *testing.T) {
 			},
 			status: http.StatusUnauthorized,
 			err:    "authentication failed: invalid credentials",
+		},
+		{
+			name: "two-factor enabled auth",
+			handler: func(w http.ResponseWriter, r *http.Request) {
+				assert.Equal(t, "POST", r.Method)
+
+				w.Header().Set("Content-Type", "application/json")
+				w.WriteHeader(http.StatusOK)
+
+				fmt.Fprint(w, fixture("auth/two_factor.json"))
+			},
+			status: http.StatusOK,
+			err:    "authentication failed: your account has two-factor authentication enabled, which is not possible to support with the API.",
 		},
 		{
 			name: "rate limit exceeded",